new_data/templates/pyproject.toml
[project]
name = "duo-talk-persona-phase1"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["pyyaml"]

[tool.pytest.ini_options]
testpaths = ["templates/tests"]
pythonpath = ["."]
